requirements.txt
# Core Neo N3 Blockchain Dependencies
neo-mamba==3.1.0
neo3-boa==1.4.1
neo3crypto==0.4.4
base58==2.1.1

# Networking & HTTP
aiohttp>=3.9.0
httpx>=0.27.0
httpx-sse>=0.4.0
requests==2.31.0
urllib3>=2.2.0

# FastAPI Backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# AI & Environment
sudo-ai>=0.1.0
python-dotenv>=1.0.0

# AWS/Storage (for IPFS and image storage)
# Pinned to 1.35.95 to avoid MissingContentLength error with 4Everland
boto3==1.35.95
botocore==1.35.99
requests-toolbelt>=1.0.0

# Image Processing
Pillow>=10.0.0

# Data Structures & Serialization
orjson>=3.10.0
jsonschema>=4.21.0
referencing>=0.35.0

# Database
sqlalchemy>=2.0.0

# Utilities
typing_extensions>=4.10.0
packaging>=24.0
